fdasf

+++ tools/train.py
@@ -124,7 +124,7 @@ def main(args, cfgs):
     if cfg.LOCAL_RANK == 0:
         os.system('cp %s %s' % (args.cfg_file, output_dir))
 
-    if cfg.LOCAL_RANK == 0:
+    if cfg.LOCAL_RANK == 0 and os.path.exists(output_dir/'tensorboard'):
 
         for file_name in os.listdir(output_dir/'tensorboard'):
             os.remove(output_dir/'tensorboard'/file_name)
